page_analyzer/app.py
from flask import Flask, render_template
from dotenv import load_dotenv
import os

load_dotenv()

app = Flask(__name__)


@app.route('/')
def index():
    return render_template('index.html')


app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
